pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "holdem-agent-bench"
version = "0.1.0"
description = "A benchmark where AI agents face off at the poker table"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "HAB Team"}]
dependencies = [
  "pokerkit>=0.5",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "httpx>=0.27",
  "typer>=0.12",
  "rich>=13.7",
  "pyyaml>=6.0",
  "pydantic>=2.6",
  "mcp>=1.0",
  "numpy>=1.26",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-cov>=4.1"]

[project.scripts]
hab = "hab.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
hab = ["presets/*.yaml", "templates/**/*"]
